long -> int

+++ bright.h
@@ -1,10 +1,10 @@
 #ifndef _BRIGHT_H
 #define _BRIGHT_H
 
-long current_brightness();
-long max_brightness();
+int current_brightness();
+int max_brightness();
 
-void set_brightness(long level);
-void inc_brightness(long amount);
+void set_brightness(int level);
+void inc_brightness(int amount);
 
 #endif // _BRIGHT_H
